invoice_backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.14
django-cors-headers>=4.3
mysqlclient>=2.2
reportlab>=4.0
Pillow>=10.0
